Remove outdated comments
The changeset removes outdated comments from the `git-se.py`
file. It deletes a block of commented-out code that was
previously used to display information about a diff. The
patch also removes a commented-out print statement related
to the path of the repository's working directory. These
comments are no longer relevant and have been removed to
clean up the code.