Introduce a new variable for unwrapped description
Introduce a new variable `pd_com_line_unwrapped` to store
the unwrapped version of the existing `pd_com_line` variable
in the `git-se.py` file. This change is made within the
`main` function at line 501. The new variable is assigned
the value of `pd_com_line`, which is stripped of leading and
trailing whitespaces.