Pressing KEY_RIGHT will select all lines of the same kind
This changeset modifies the `git-se.py` file. It introduces
a new functionality where pressing the `KEY_RIGHT` key will
now select all lines of the same kind as the currently
selected line. It iterates through the lines below the
current line until it finds a line that is not of the same
kind, toggling the selection status of each line
accordingly. This allows for selecting multiple lines of the
same type with a single key press.