Allow splitting patches also for ADDED files
This changeset modifies the `git-se.py` file to allow
splitting changesets for files that are added. Previously,
the condition only checked for `DeltaStatus.MODIFIED`
status, but now it has been updated to also include
`DeltaStatus.ADDED`. This change ensures that when a file is
added, it will be considered for splitting changesets.