Do not output binary patch to nano template
The changeset modifies the `export_patch` function in the
`git-se.py` file. It adds a condition to check if the patch
is binary before writing it to the file. If the patch is
binary, it will not be output to the nano template. This
change ensures that binary patches are not included in the
output.