Run `nano` for user to edit the final description
The changeset modifies the `git-se.py` file to include a new subprocess call to
run the `nano` text editor for the user to edit the final description. This
subprocess call opens the `git-se._stage_desc.txt` file using `nano`. This
allows the user to make changes to the description before proceeding further in
the script.